sahayak-main/requirements.txt
google-genai
google-generativeai 
Pillow
streamlit
fpdf
dotenv
pyaudio




sounddevice 
numpy 
scipy 
faster-whisper 


